feat: complete project restructuring with domain-specific utilities (Phase D)
- Split lib/utils.ts into domain-specific modules:
  * lib/utils/ui.ts - UI utilities (cn function)
  * lib/utils/formatting.ts - Date/text formatting utilities
  * lib/utils/blog.ts - Blog-specific utilities (sorting, tags, stats)
- Maintain 100% backward compatibility via re-exports in lib/utils.ts
- Enable organized imports: @/lib/utils/ui, @/lib/utils/blog, etc.
- Update documentation with new feature-based architecture:
  * CHANGELOG.md - Added Phase 4 restructuring details
  * CLAUDE.md - Added project architecture section with import patterns
  * README.md - Updated project structure with feature-based organization
  * RESTRUCTURING_PLAN.md - Marked as completed with implementation date
- Zero breaking changes - all existing imports continue working
- Build verified: 104 static pages generated successfully

Phase 4 complete: Codebase restructured to 2025 best practices with domain-driven design.
