+++ lib/utils/formatting.ts
@@ -1,0 +1,8 @@
+export function formatDate(input: string | number): string {
+  const date = new Date(input);
+  return date.toLocaleDateString("en-US", {
+    month: "long",
+    day: "numeric", 
+    year: "numeric",
+  });
+}
